pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "livermri-crossseq"
version = "0.1.0"
description = "Liver MRI cross-sequence nnUNet orchestration and experiment tooling."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "nibabel",
  "SimpleITK",
  "numpy",
  "pandas",
  "pyyaml",
  "scikit-learn",
  "scipy",
  "matplotlib",
  "tqdm",
  "jinja2",
  "openpyxl",
]

[project.scripts]
livermri-bootstrap = "livermri_crossseq.cli.bootstrap_workspace:main"
livermri-plan-anchor = "livermri_crossseq.cli.plan_anchor_pilot:main"
livermri-export-crossseq = "livermri_crossseq.cli.export_registered_multibranch_dataset:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
